+++ src/components/functions/Copy.jsx
@@ -93,7 +93,7 @@ export function Copy({ currentContext, isDisabled, onStatusUpdate }) {
           {!longPressDisabled && <span className='italic'>Hold for more options</span>}
         </Tooltip.Content>
       </Tooltip>
-      <Dropdown.Popover className='w-fit min-w-48' placement='bottom left'>
+      <Dropdown.Popover className='w-fit min-w-60' placement='bottom left'>
         <Dropdown.Menu onAction={handleAction}>
           {dropdownItems.map((item) => (
             <Dropdown.Item id={item.id} key={item.id} textValue={item.label}>